chore: apply feedback
