+++ packages/stargate/src/modules/authz/queries.spec.ts
@@ -3,7 +3,8 @@ import { Tendermint34Client } from "@cosmjs/tendermint-rpc";
 import { assertDefined, sleep } from "@cosmjs/utils";
 import { GenericAuthorization } from "cosmjs-types/cosmos/authz/v1beta1/authz";
 
-import { makeLinkPath, QueryClient } from "../../queryclient";
+import { makeLinkPath } from "../../paths";
+import { QueryClient } from "../../queryclient";
 import { SigningStargateClient } from "../../signingstargateclient";
 import { assertIsDeliverTxSuccess } from "../../stargateclient";
 import {
@@ -41,7 +42,7 @@ xdescribe("AuthzExtension", () => {
         // Use address 1 and 2 instead of 0 to avoid conflicts with other delegation tests
         // This must match `voterAddress` above.
         hdPaths: [makeLinkPath(1), makeLinkPath(2)],
-        prefix: "link",
+        prefix: simapp.prefix,
       });
       const client = await SigningStargateClient.connectWithSigner(
         simapp.tendermintUrl,
